requirements.txt
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-ai-contentsafety>=1.0.0
mlflow>=2.8.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
pyyaml>=6.0.0
azure-ai-ml
azure-identity
